requirements.txt
datasetinsights~=1.1.1
setuptools~=58.0.4
tensorflow~=2.7.0
six~=1.16.0
numpy~=1.20.3
contextlib2~=21.6.0
pillow~=8.4.0
scipy~=1.7.2
ipython~=7.29.0
google~=3.0.0
matplotlib~=3.4.3
pycocotools~=2.0.3
lvis~=0.5.3
pandas~=1.3.4
lxml~=4.6.4
opencv-python~=4.5.4.60
scikit-learn~=1.0.1